// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type ActivityEventPage = {
  items: Array<ActivityEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type Project = {
  activityEvents: Maybe<ActivityEventPage>;
  addNftTierEvents: Maybe<AddNftTierEventPage>;
  addToBalanceEvents: Maybe<AddToBalanceEventPage>;
  autoIssueEvents: Maybe<AutoIssueEventPage>;
  balance: Scalars['BigInt']['output'];
  borrowLoanEvents: Maybe<BorrowLoanEventPage>;
  bridgeClaimEvents: Maybe<BridgeClaimEventPage>;
  burnEvents: Maybe<BurnEventPage>;
  buybackPoolEvents: Maybe<BuybackPoolEventPage>;
  cashOutTokensEvents: Maybe<CashOutTokensEventPage>;
  chainId: Scalars['Int']['output'];
  contributorsCount: Scalars['Int']['output'];
  coverImageUri: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Int']['output'];
  createdWithinTrendingWindow: Maybe<Scalars['Boolean']['output']>;
  creator: Scalars['String']['output'];
  currency: Maybe<Scalars['BigInt']['output']>;
  decimals: Maybe<Scalars['Int']['output']>;
  deployErc20Events: Maybe<DeployErc20EventPage>;
  deployer: Scalars['String']['output'];
  description: Maybe<Scalars['String']['output']>;
  discord: Maybe<Scalars['String']['output']>;
  domain: Maybe<Scalars['String']['output']>;
  farcaster: Maybe<Scalars['String']['output']>;
  handle: Maybe<Scalars['String']['output']>;
  id: Scalars['String']['output'];
  infoUri: Maybe<Scalars['String']['output']>;
  isRevnet: Maybe<Scalars['Boolean']['output']>;
  liquidateLoanEvents: Maybe<LiquidateLoanEventPage>;
  logoUri: Maybe<Scalars['String']['output']>;
  manualBurnEvents: Maybe<ManualBurnEventPage>;
  manualMintTokensEvents: Maybe<ManualMintTokensEventPage>;
  metadata: Maybe<Scalars['JSON']['output']>;
  metadataUri: Maybe<Scalars['String']['output']>;
  mintNftEvents: Maybe<MintNftEventPage>;
  mintTokensEvents: Maybe<MintTokensEventPage>;
  name: Maybe<Scalars['String']['output']>;
  nftHooks: Maybe<NftHookPage>;
  nfts: Maybe<NftPage>;
  nftsMintedCount: Scalars['Int']['output'];
  operatorPermissionsSetEvents: Maybe<OperatorPermissionsSetEventPage>;
  owner: Scalars['String']['output'];
  participants: Maybe<ParticipantPage>;
  payDisclosure: Maybe<Scalars['String']['output']>;
  payEvents: Maybe<PayEventPage>;
  paymentsCount: Scalars['Int']['output'];
  permissionHolders: Maybe<PermissionHolderPage>;
  projectCreateEvents: Maybe<ProjectCreateEventPage>;
  projectId: Scalars['Int']['output'];
  projectMoments: Maybe<ProjectMomentPage>;
  projectPayers: Maybe<ProjectPayerPage>;
  projectTagline: Maybe<Scalars['String']['output']>;
  projectTransferEvents: Maybe<ProjectTransferEventPage>;
  reallocateLoanEvents: Maybe<ReallocateLoanEventPage>;
  redeemCount: Scalars['Int']['output'];
  redeemVolume: Scalars['BigInt']['output'];
  redeemVolumeUsd: Scalars['BigInt']['output'];
  removeNftTierEvents: Maybe<RemoveNftTierEventPage>;
  repayLoanEvents: Maybe<RepayLoanEventPage>;
  reservedTokenSupply: Scalars['BigInt']['output'];
  rulesetQueuedEvents: Maybe<RulesetQueuedEventPage>;
  sendPayoutToSplitEvents: Maybe<SendPayoutToSplitEventPage>;
  sendPayoutsEvents: Maybe<SendPayoutsEventPage>;
  sendReservedTokensToSplitEvents: Maybe<SendReservedTokensToSplitEventPage>;
  sendReservedTokensToSplitsEvents: Maybe<SendReservedTokensToSplitsEventPage>;
  setUriEvents: Maybe<SetUriEventPage>;
  storeAutoIssuanceAmountEvent: Maybe<StoreAutoIssuanceAmountEventPage>;
  suckerGroup: Maybe<SuckerGroup>;
  suckerGroupId: Scalars['String']['output'];
  swapEvents: Maybe<SwapEventPage>;
  tags: Maybe<Array<Scalars['String']['output']>>;
  telegram: Maybe<Scalars['String']['output']>;
  token: Maybe<Scalars['String']['output']>;
  tokenSupply: Scalars['BigInt']['output'];
  tokenSymbol: Maybe<Scalars['String']['output']>;
  tokens: Maybe<Array<Scalars['String']['output']>>;
  trendingPaymentsCount: Scalars['Int']['output'];
  trendingScore: Scalars['BigInt']['output'];
  trendingVolume: Scalars['BigInt']['output'];
  trendingVolumeUsd: Scalars['BigInt']['output'];
  twitter: Maybe<Scalars['String']['output']>;
  useAllowanceEvents: Maybe<UseAllowanceEventPage>;
  version: Scalars['Int']['output'];
  volume: Scalars['BigInt']['output'];
  volumeUsd: Scalars['BigInt']['output'];
};
// ...
export type ParticipantsQueryVariables = Exact<{
  where?: InputMaybe<ParticipantFilter>;
  orderBy?: InputMaybe<Scalars['String']['input']>;
  orderDirection?: InputMaybe<Scalars['String']['input']>;
  limit?: InputMaybe<Scalars['Int']['input']>;
}>;


export type ParticipantsQuery = { participants: { totalCount: number, items: Array<{ chainId: number, address: string, volume: any, lastPaidTimestamp: number, balance: any, erc20Balance: any, creditBalance: any }> } };

export type HasPermissionQueryVariables = Exact<{
  account: Scalars['String']['input'];
  chainId: Scalars['Float']['input'];
  projectId: Scalars['Float']['input'];
  operator: Scalars['String']['input'];
  version: Scalars['Float']['input'];
}>;


export type HasPermissionQuery = { permissionHolder: { permissions: Array<number> | null } | null };

export type ProjectQueryVariables = Exact<{
  projectId: Scalars['Float']['input'];
  chainId: Scalars['Float']['input'];
  version: Scalars['Float']['input'];
}>;


export type ProjectQuery = { project: { projectId: number, metadataUri: string | null, handle: string | null, createdAt: number, suckerGroupId: string, logoUri: string | null, name: string | null, version: number, token: string | null, decimals: number | null, currency: any | null, tokenSymbol: string | null, isRevnet: boolean | null } | null };
// ...
export type SuckerGroupQueryVariables = Exact<{
  id: Scalars['String']['input'];
}>;


export type SuckerGroupQuery = { suckerGroup: { id: string, tokenSupply: any, projects: { items: Array<{ balance: any, chainId: number, currency: any | null, decimals: number | null, projectId: number, token: string | null, tokenSupply: any, tokenSymbol: string | null, version: number, suckerGroupId: string }> } | null } | null };
// ...
export const ParticipantsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"Participants"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"where"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"participantFilter"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"orderBy"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"orderDirection"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"limit"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"participants"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"where"},"value":{"kind":"Variable","name":{"kind":"Name","value":"where"}}},{"kind":"Argument","name":{"kind":"Name","value":"orderBy"},"value":{"kind":"Variable","name":{"kind":"Name","value":"orderBy"}}},{"kind":"Argument","name":{"kind":"Name","value":"orderDirection"},"value":{"kind":"Variable","name":{"kind":"Name","value":"orderDirection"}}},{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"Variable","name":{"kind":"Name","value":"limit"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"totalCount"}},{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"chainId"}},{"kind":"Field","name":{"kind":"Name","value":"address"}},{"kind":"Field","name":{"kind":"Name","value":"volume"}},{"kind":"Field","name":{"kind":"Name","value":"lastPaidTimestamp"}},{"kind":"Field","name":{"kind":"Name","value":"balance"}},{"kind":"Field","name":{"kind":"Name","value":"erc20Balance"}},{"kind":"Field","name":{"kind":"Name","value":"creditBalance"}}]}}]}}]}}]} as unknown as DocumentNode<ParticipantsQuery, ParticipantsQueryVariables>;
// ...
export const ProjectDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"Project"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"projectId"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Float"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"chainId"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Float"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"version"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Float"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"project"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"projectId"},"value":{"kind":"Variable","name":{"kind":"Name","value":"projectId"}}},{"kind":"Argument","name":{"kind":"Name","value":"chainId"},"value":{"kind":"Variable","name":{"kind":"Name","value":"chainId"}}},{"kind":"Argument","name":{"kind":"Name","value":"version"},"value":{"kind":"Variable","name":{"kind":"Name","value":"version"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"projectId"}},{"kind":"Field","name":{"kind":"Name","value":"metadataUri"}},{"kind":"Field","name":{"kind":"Name","value":"handle"}},{"kind":"Field","name":{"kind":"Name","value":"createdAt"}},{"kind":"Field","name":{"kind":"Name","value":"suckerGroupId"}},{"kind":"Field","name":{"kind":"Name","value":"logoUri"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"version"}},{"kind":"Field","name":{"kind":"Name","value":"token"}},{"kind":"Field","name":{"kind":"Name","value":"decimals"}},{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"tokenSymbol"}},{"kind":"Field","name":{"kind":"Name","value":"isRevnet"}}]}}]}}]} as unknown as DocumentNode<ProjectQuery, ProjectQueryVariables>;
// ...
export const SuckerGroupDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"SuckerGroup"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"suckerGroup"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"tokenSupply"}},{"kind":"Field","name":{"kind":"Name","value":"projects"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"balance"}},{"kind":"Field","name":{"kind":"Name","value":"chainId"}},{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"decimals"}},{"kind":"Field","name":{"kind":"Name","value":"projectId"}},{"kind":"Field","name":{"kind":"Name","value":"token"}},{"kind":"Field","name":{"kind":"Name","value":"tokenSupply"}},{"kind":"Field","name":{"kind":"Name","value":"tokenSymbol"}},{"kind":"Field","name":{"kind":"Name","value":"version"}},{"kind":"Field","name":{"kind":"Name","value":"suckerGroupId"}}]}}]}}]}}]}}]} as unknown as DocumentNode<SuckerGroupQuery, SuckerGroupQueryVariables>;